% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{CATE_standardize}
\alias{CATE_standardize}
\title{Standardizes a point of interest and a sample
to improve comparability between dimensions (for CATE Setup)}
\usage{
CATE_standardize(x, data)
}
\arguments{
\item{x}{Point of interest}

\item{data}{Sample under Consideration}
}
\value{
The sample sorted relative to x
}
\description{
Standardizes a point of interest and a sample
to improve comparability between dimensions (for CATE Setup)
}
