% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_gen.R
\name{CATE_test_data}
\alias{CATE_test_data}
\title{Generate a data set for the nonparametric regression setup.}
\usage{
CATE_test_data(n_obs, cov_dim)
}
\arguments{
\item{n_obs}{Number of desired observations}

\item{cov_dim}{Dimensionality of covariates}
}
\value{
A suitable matrix of data.
}
\description{
Generate a data set for the nonparametric regression setup.
}
