% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preChecks.R
\name{data_check}
\alias{data_check}
\title{Check the provided data for correct format etc}
\usage{
data_check(data)
}
\arguments{
\item{data}{The data set containing the observations.}
}
\value{
TRUE or FALSE
}
\description{
Check the provided data for correct format etc
}
