% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preChecks.R
\name{data_point_compatibility_check}
\alias{data_point_compatibility_check}
\title{Check the provided data and point of interest for compatibility}
\usage{
data_point_compatibility_check(x, data)
}
\arguments{
\item{x}{The point of interest}

\item{data}{The data set containing the observations.}
}
\value{
TRUE or FALSE
}
\description{
Check the provided data and point of interest for compatibility
}
