% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{dist}
\alias{dist}
\title{Calculate the distance between two points of interest}
\usage{
dist(x, y)
}
\arguments{
\item{x}{Point A of interest}

\item{y}{Point B of interest}
}
\value{
The Euclidean distance between point A and B
}
\description{
Calculate the distance between two points of interest
}
