% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNN_DML_CATE.R
\name{DNN_DML2}
\alias{DNN_DML2}
\title{Calculate the DNN-DML2 CATE-Estimator}
\usage{
DNN_DML2(
  x,
  data,
  s,
  n_folds,
  presorted = FALSE,
  standardize = FALSE,
  asymp_approx_weights = TRUE
)
}
\arguments{
\item{x}{The point of interest.}

\item{data}{The data set containing the observations.
As a matrix containing the response value in the first column
and each covariate in a subsequent columns}

\item{s}{The subsampling scale.}

\item{n_folds}{The number of folds used for crossfitting}

\item{presorted}{True or False whether the data is sorted according to its
distance to the point of interest (default value = FALSE)}

\item{standardize}{True or False whether to standardize (default value = FALSE)}

\item{asymp_approx_weights}{True or False whether to use an asymptotic
approximation for the weights (default value = FALSE)}
}
\value{
A number.
}
\description{
Calculate the DNN-DML2 CATE-Estimator
}
