% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNN_LOO_CATE_mult.R
\name{DNN_LOO_mult}
\alias{DNN_LOO_mult}
\title{Calculate the DNN-DML2 CATE-Estimator for a list of points of interest}
\usage{
DNN_LOO_mult(points, data, s, asymp_approx_weights = TRUE)
}
\arguments{
\item{points}{A matrix containing points of interest as its rows.}

\item{data}{The data set containing the observations.
As a matrix containing the response value in the first column
and each covariate in a subsequent columns}

\item{s}{The subsampling scale.}

\item{asymp_approx_weights}{True or False whether to use an asymptotic
approximation for the weights (default value = FALSE)}
}
\value{
A number.
}
\description{
Calculate the DNN-DML2 CATE-Estimator for a list of points of interest
}
