% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNN_DML_Nuisance.R
\name{DNN_Nuisance}
\alias{DNN_Nuisance}
\title{This function estimates the propensity score, mu0 and mu1
at a given point}
\usage{
DNN_Nuisance(x, data, s, asymp_approx_weights = TRUE)
}
\arguments{
\item{x}{Point of interest}

\item{data}{Data matrix for the CATE setting}

\item{s}{Subsampling Scale}

\item{asymp_approx_weights}{True or False whether to use an asymptotic
approximation for the weights (default value = TRUE)}
}
\description{
This function estimates the propensity score, mu0 and mu1
at a given point
}
