% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{generate_folds}
\alias{generate_folds}
\title{Given a number of observations and a number of desired folds,
returns the a list of vectors containing the indices of each fold}
\usage{
generate_folds(n_obs, n_folds)
}
\arguments{
\item{n_obs}{Number of Observations}

\item{n_folds}{Number of desired folds}
}
\description{
Given a number of observations and a number of desired folds,
returns the a list of vectors containing the indices of each fold
}
