% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_CATE.R
\name{naive_CATE_TDNN}
\alias{naive_CATE_TDNN}
\title{Naive estimator for conditional average treatment effects based on the TDNN
estimator.}
\usage{
naive_CATE_TDNN(
  x,
  data,
  s1,
  s2,
  presorted = FALSE,
  standardize = FALSE,
  asymp_approx_weights = TRUE
)
}
\arguments{
\item{x}{The point of interest.}

\item{data}{The data set containing the observations.
As a matrix containing the response value in the first column, the
treatment status in the second and each covariate in subsequent columns}

\item{s1}{The first subsampling scale.}

\item{s2}{The second subsampling scale.}

\item{presorted}{True or False whether the data is sorted according to its
distance to the point of interest (default value = FALSE)}

\item{standardize}{True or False whether to standardize (default value = FALSE)}

\item{asymp_approx_weights}{True or False whether to use an asymptotic
approximation for the weights (default value = FALSE)}
}
\value{
A number.
}
\description{
Naive estimator for conditional average treatment effects based on the TDNN
estimator.
}
