% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preChecks.R
\name{point_check}
\alias{point_check}
\title{Check the provided point of interest for correct format etc}
\usage{
point_check(x)
}
\arguments{
\item{x}{The point of interest}
}
\value{
TRUE or FALSE
}
\description{
Check the provided point of interest for correct format etc
}
