% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{pre_sort}
\alias{pre_sort}
\title{Sorts a sample according to its distance relative
to a point of interest}
\usage{
pre_sort(x, data)
}
\arguments{
\item{x}{Point of interest}

\item{data}{Sample under Consideration}
}
\value{
The sample sorted relative to x
}
\description{
Sorts a sample according to its distance relative
to a point of interest
}
