% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preChecks.R
\name{samplingscale_check}
\alias{samplingscale_check}
\title{Check the provided data and sampling scale of interest for compatibility}
\usage{
samplingscale_check(s, data)
}
\arguments{
\item{s}{The subsampling scale}

\item{data}{The data set containing the observations.}
}
\value{
TRUE or FALSE
}
\description{
Check the provided data and sampling scale of interest for compatibility
}
