% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preChecks.R
\name{samplingscale_check2}
\alias{samplingscale_check2}
\title{Check the provided data and subsampling scales for compatibility}
\usage{
samplingscale_check2(s1, s2, data)
}
\arguments{
\item{s1}{The first subsampling scale}

\item{s2}{The second subsampling scale}

\item{data}{The data set containing the observations.}
}
\value{
TRUE or FALSE
}
\description{
Check the provided data and subsampling scales for compatibility
}
