% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDNN_DML_Nuisance.R
\name{TDNN_Nuisance}
\alias{TDNN_Nuisance}
\title{This function estimates the propensity score, mu0 and mu1
at a given point}
\usage{
TDNN_Nuisance(x, data, s1, s2, asymp_approx_weights = TRUE)
}
\arguments{
\item{x}{Point of interest}

\item{data}{Data matrix for the CATE setting}

\item{s1}{Subsampling Scale 1}

\item{s2}{Subsampling Scale 2}

\item{asymp_approx_weights}{True or False whether to use an asymptotic
approximation for the weights (default value = FALSE)}
}
\description{
This function estimates the propensity score, mu0 and mu1
at a given point
}
