% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNN.R
\name{DNN}
\alias{DNN}
\title{Calculate the distributional nearest neighbor estimate for a point x
based on a sample data with subsampling scale s.
The implementation is based on the equivalent representation as an L-statistic
described in Steele (2009)}
\usage{
DNN(x, data, s, presorted = FALSE)
}
\arguments{
\item{x}{The point of interest.}

\item{data}{The data set containing the observations.
As a matrix containing the response value in the first column.}

\item{s}{The subsampling scale.}
}
\value{
A number.
}
\description{
Calculate the distributional nearest neighbor estimate for a point x
based on a sample data with subsampling scale s.
The implementation is based on the equivalent representation as an L-statistic
described in Steele (2009)
}
