% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNN.R
\name{DNN}
\alias{DNN}
\title{Calculate the distributional nearest neighbor estimate for a point x
based on a sample data with subsampling scale s.
The implementation is based on the equivalent representation as an L-statistic
described in Steele (2009)}
\usage{
DNN(
  x,
  data,
  s,
  presorted = FALSE,
  standardize = FALSE,
  asymp_approx_weights = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{The point of interest.}

\item{data}{The data set containing the observations.
As a matrix containing the response value in the first column
and each covariate in a subsequent columns}

\item{s}{The subsampling scale.}

\item{presorted}{True or False whether the data is sorted according to its
distance to the point of interest (default value = FALSE)}

\item{standardize}{True or False whether to standardize (default value = FALSE)}

\item{asymp_approx_weights}{True or False whether to use an asymptotic
approximation for the weights (default value = FALSE)}
}
\value{
A number.
}
\description{
Calculate the distributional nearest neighbor estimate for a point x
based on a sample data with subsampling scale s.
The implementation is based on the equivalent representation as an L-statistic
described in Steele (2009)
}
