% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDNN.R
\name{TDNN}
\alias{TDNN}
\title{Calculate the distributional nearest neighbor estimate for a point x
based on a sample data with subsampling scale s.
The implementation is based on the equivalent representation as an L-statistic
described in Steele (2009) and the TDNN estimator from Demirkaya et al. (2024)}
\usage{
TDNN(x, data, s1, s2)
}
\arguments{
\item{x}{The point of interest.}

\item{data}{The data set containing the observations.}

\item{s1}{The first subsampling scale.}

\item{s2}{The second subsampling scale.}
}
\value{
A number.
}
\description{
Calculate the distributional nearest neighbor estimate for a point x
based on a sample data with subsampling scale s.
The implementation is based on the equivalent representation as an L-statistic
described in Steele (2009) and the TDNN estimator from Demirkaya et al. (2024)
}
