% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDNN.R
\name{TDNN}
\alias{TDNN}
\title{Calculate the distributional nearest neighbor estimate for a point x
based on a sample data with subsampling scale s.
The implementation is based on the equivalent representation as an L-statistic
described in Steele (2009) and the TDNN estimator from Demirkaya et al. (2024)}
\usage{
TDNN(
  x,
  data,
  s1,
  s2,
  presorted = FALSE,
  standardize = FALSE,
  asymp_approx_weights = TRUE
)
}
\arguments{
\item{x}{The point of interest.}

\item{data}{The data set containing the observations.
As a matrix containing the response value in the first column
and each covariate in a subsequent columns}

\item{s1}{The first subsampling scale.}

\item{s2}{The second subsampling scale.}

\item{presorted}{True or False whether the data is sorted according to its
distance to the point of interest (default value = FALSE)}

\item{standardize}{True or False whether to standardize (default value = FALSE)}

\item{asymp_approx_weights}{True or False whether to use an asymptotic
approximation for the weights (default value = FALSE)}
}
\value{
A number.
}
\description{
Calculate the distributional nearest neighbor estimate for a point x
based on a sample data with subsampling scale s.
The implementation is based on the equivalent representation as an L-statistic
described in Steele (2009) and the TDNN estimator from Demirkaya et al. (2024)
}
